[Fix] Invalid deletion request
- 회의록 작성 중이 아닐 때, 컴포넌트가 destory되면 undefined 문서를 삭제 시도하는 버그 수정

+++ frontend/src/components/MeetingMinuteEditor.vue
@@ -80,10 +80,12 @@ export default {
   //   window.removeEventListener('beforeunload', this.leave);
   // },
   async beforeDestroy() {
+    const meetingMinuteId = sessionStorage.getItem('meetingMinuteId');
+    if (meetingMinuteId == null) return;
+    
     if ((this.title == null || this.title == "")
       && (this.date == null || this.date == "")
       && (this.place == null || this.place == "")) {
-      const meetingMinuteId = sessionStorage.getItem('meetingMinuteId');
       await axios.delete(`/api/meeting?board_id=${this.board_id}&meetingMinuteId=${meetingMinuteId}`, {
         headers: this.$store.getters.headers
       })
